+++ res/views/components/button.blade.php
@@ -1,5 +1,11 @@
 @isset($href)
-    @tag('a', $attributes->merge(compact('href')), $slot)
+    @tag('a', $attributes->merge(compact('href')))
+        <s-text :value="$text">{{ $slot }}</s-text>
+    @endtag
+@elseif(isset($text) || $slot->isNotEmpty())
+    @tag('button', $attributes->merge(compact('type')))
+        <s-text :value="$text">{{ $slot }}</s-text>
+    @endtag
 @else
-    @tag('button', $attributes->merge(compact('type')), $slot)
+    @tag('input', $attributes->merge(compact('type', 'name', 'value')))
 @endisset