+++ src/Components/Button.php
@@ -20,6 +20,8 @@ class Button extends BladeComponent
         string $href = null,
         string $route = null,
         string|array $parameters = [],
+        public ?string $name = null,
+        public ?string $value = null,
         public ?string $text = null,
     ) {
         $type = strtolower($type);