+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.5.0...v0.5.1](https://github.com/sikessem/ui/compare/v0.5.0...v0.5.1)
 
+- [0.x] Add [button](https://github.com/sikessem/ui#button-component) types
+
 ## [v0.5.1](https://github.com/sikessem/ui/releases/tag/v0.5.1) - 2023-08-02
 
 - [0.x] Improve plugins architecture